requirements.txt
flask==2.3.2
python-telegram-bot==2.0
flask-limiter==3.5.0
python-dotenv==1.0.0